% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb_scan_interface.R
\name{call_scan}
\alias{call_scan}
\alias{import_from_dir}
\alias{scan_import}
\title{CWB-Scan interface for R}
\usage{
call_scan(dir_path, parameters)

import_from_dir(dir_path, col_names)

scan_import(dir_path, parameters, col_names = names(parameters))
}
\arguments{
\item{dir_path}{Name of directory to save results in}

\item{parameters}{list of character vectors with corpora, p_attributes and
s_attributes}

\item{col_names}{column names for the parameters. by default read from the
names of the list provided in parameters}
}
\description{
A convenience wrapper for scripting with cwb-scan. Creates data with
cwb-scan-corpus. \code{scan_import()} additionally imports the entire directory
with fread as data.table using \code{fread()} and \code{rbindlist()}.
}
\examples{
\dontrun{
lol <- list(
  corpus = c("BASE", "BROWN", "FROWN"),
  p_attr = c("word", "lemma"),
  s_attr = c("text_id")
)

full <- scan_import("data/", lol)
full <- import_from_dir("data/", names(lol))
call_scan("data/", lol)
}
}
