% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb_scan_interface.R
\docType{data}
\name{corpus}
\alias{corpus}
\title{Create a data grid}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
corpus
}
\arguments{
\item{parameters}{Named list of parameters for "corpus", "s_attr", "p_attr"}
}
\description{
Create a grid with combinations of parameters to scan. This
includes a look-up table to replace alternative attribute names. Depends on
installation
}
\keyword{datasets}
