% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb_decode.R
\name{cwb_decode}
\alias{cwb_decode}
\title{Decode entire corpus}
\usage{
cwb_decode(corpus, p_attrs, s_attrs)
}
\arguments{
\item{corpus}{character name of the corpus in the cwb registry}

\item{p_attrs}{character vector with p-attributes}

\item{s_attrs}{character vector with s-attributes}
}
\description{
A convenience wrapper generating \code{system()} calls to cwb-decode
Requires a working installation of cwb
}
\examples{
\dontrun{
cmd <- cwb_decode("BROWN", c("word", "lemma"), c("text_id", "text_genre"))
stopifnot(is.character(cmd), length(cmd) == 1)
}

}
