% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb_scan_interface.R
\name{cwb_scan}
\alias{cwb_scan}
\title{Call cwb-scan-corpus with vectors of parameters}
\usage{
cwb_scan(corpus, p_attr, s_attr, dir_path = tempdir(), constraint = NULL)
}
\arguments{
\item{corpus}{Name of the encoded cwb corpus.}

\item{p_attr}{positional attribute to scan}

\item{s_attr}{structural attribute to scan}

\item{dir_path}{Name of the directory to save files in}

\item{constraint}{additional character string with constraints, see \verb{man cwb-scan-corpus}}
}
\description{
This calls cwb-scan-corpus with parameters provided by vectors for all their
combinations. Currently one p_attribute and one s_attribute are hardcoded.
}
