% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwbwrapr-package.R
\docType{package}
\name{cwbwrapr-package}
\alias{cwbwrapr}
\alias{cwbwrapr-package}
\title{cwbwrapr: A Selection of Convenience Tools to Call CWB and CQP Commands from R}
\description{
This package is used to scan and query corpus data from an existing CWB installation. Basically just system calls, no convoluted object classes or additional syntax to mirror already existing CWB functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{www.alexraw.xyz}
}

}
\author{
\strong{Maintainer}: Alexander Rauhut \email{alexraw@mailbox.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
