% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb_scan_interface.R
\name{get_data_grid}
\alias{get_data_grid}
\title{Create a data grid}
\usage{
get_data_grid(parameters)
}
\arguments{
\item{parameters}{Named list of parameters for "corpus", "s_attr", "p_attr"}
}
\description{
Create a grid with combinations of parameters to scan. This
includes a look-up table to replace alternative attribute names. Depends on
installation
}
