% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqp_interface.R
\name{query_cqp}
\alias{query_cqp}
\title{Query CQP from R}
\usage{
query_cqp(query, corpus = NULL, n = 5, path = NULL)
}
\arguments{
\item{query}{character string with the cqp commands to be used
if multiple commands are stringed together, \verb{;} has to be used as separator}

\item{corpus}{character optionally provide corpus to be used}

\item{n}{integer left and right context}

\item{path}{name of the file to save the results to}
}
\description{
A convenience wrapper generating \code{system()} calls to cqp -c
Requires a working installation of cwb
}
\examples{
\dontrun{
cqp_query <- '"example"'
query_cqp(cqp_query, "BROWN")
}
}
