% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_freqs}
\alias{read_freqs}
\alias{write_freqs}
\alias{fread_freqs}
\alias{fwrite_freqs}
\title{Read and write plain text frequency lists}
\usage{
read_freqs(
  file,
  header = FALSE,
  cols = list(0L, ""),
  sep = "\\t",
  comment.char = "",
  na.strings = "",
  quote = "",
  allowEscapes = FALSE,
  nlines = sh_count_lines(file),
  ...
)

write_freqs(..., sep = "\\t", quote = FALSE, na = "")

fread_freqs(
  ...,
  header = FALSE,
  sep = "\\t",
  quote = "",
  na.strings = NULL,
  stringsAsFactors = FALSE
)

fwrite_freqs(..., sep = "\\t", sep2 = " ", quote = FALSE)
}
\arguments{
\item{header}{logical. whether or not the first line should be used as column
names. Note: \code{names} in cols take precedence.}

\item{cols}{list. a list with column types (e.g. \code{0}, for numeric,
\code{""} for character), if named, the names will be used as column/vector names.
By default, column 1 is expected to contain integer frequencies and column 2
strings with types}

\item{sep}{character.}

\item{comment.char}{character.}

\item{na.strings}{character.}

\item{quote}{character.}

\item{nlines}{integer. number of lines in file, see \code{nlines} in \code{scan()}}

\item{...}{further arguments to be passed to \code{scan}}

\item{.x}{character. path to file or connection, see \code{scan()}}

\item{skip}{integer. how many lines to skip, see \code{scan()}}
}
\value{
data.frame or data.table
}
\description{
Convenience wrapper to import typical output from CWB tools.
The expected file format is tab delimited, only containing data, no quotes,
missing values, comments or header.
}
\details{
These are convenience wrappers around \code{scan} or \code{fread} with sane
defaults for common frequency list formats.
In \code{read_freqs}, \code{wc -l} is run if available to pass the line number to scan
This reduces memory overhead substantially and can also be a bit faster.
}
\examples{
\dontrun{
path <- "brown_word_per_id.txt"
out <- read_freqs(path, list(f = 0L, type = "", text_id = ""))
}
}
\seealso{
\code{\link[=scan(), fread]{scan(), fread()}}
}
