% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_vrt}
\alias{read_vrt}
\alias{read_wpl}
\alias{write_vrt}
\alias{write_wpl}
\title{Read and write VRT or WPL files}
\usage{
read_vrt(file, no_xml = FALSE)

read_wpl(file, no_xml = FALSE)

write_vrt(file)

write_wpl(file)
}
\arguments{
\item{.x}{character. path to file or connection, see \code{scan()}}

\item{header}{logical. whether or not the first line should be used as column
names. Note: \code{names} in cols take precedence.}

\item{cols}{list. a list with column types (e.g. \code{0}, for numeric,
\code{""} for character), if named, the names will be used as column/vector names.
By default, column 1 is expected to contain integer frequencies and column 2
strings with types}

\item{sep}{character. see \code{scan}}

\item{comment.char}{character. off by default. replace if necessary}

\item{quote}{character. see \code{scan}}

\item{na.strings}{character. see \code{scan}}

\item{nlines}{integer. number of lines in file, see \code{nlines} in \code{scan()}}

\item{...}{further arguments to be passed to \code{scan}}
}
\value{
data.frame
}
\description{
Import VRT (verticalized text) or WLP (word per line) files
}
\details{
These formats are used by common corpus indexing software, like CWB
(.vrt) and Sketchengine (.wpl). These functions are convenience wrappers with
to read from and create such file formats.
}
\seealso{
\code{\link[=read.table]{read.table()}}
}
