% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{read_freqs}
\alias{read_freqs}
\alias{fread_freqs}
\title{Import plain text frequency lists}
\usage{
read_freqs(
  path,
  cols = list(f = 0L, type = ""),
  df = TRUE,
  n = sh_count_lines(path),
  comment.char = "",
  quiet = TRUE,
  ...
)

fread_freqs(
  ...,
  header = FALSE,
  sep = "\\t",
  quote = "",
  na.strings = NULL,
  stringsAsFactors = FALSE
)
}
\arguments{
\item{path}{character. path to file}

\item{cols}{list. a list with column types (e.g. \code{0}, for numeric,
\code{""} for character), if named, the names will be used as column/vector names.
By default, column 1 is expected to contain integer frequencies and column 2
strings with types}

\item{df}{boolean. If TRUE (default) a \code{data.frame} is returned.}

\item{n}{number of lines in file, see \code{nlines} in \code{scan()}}

\item{comment.char}{character. off by default. replace if necessary}

\item{quiet}{boolean. If TRUE, number of records is printed after import}

\item{...}{further arguments to be passed to \code{scan} or \code{fread}}

\item{header}{see \code{scan} or \code{fread}}

\item{sep}{see \code{scan} or \code{fread}}

\item{quote}{see \code{scan} or \code{fread}}

\item{na.strings}{see \code{scan} or \code{fread}}

\item{stringsAsFactors}{see \code{scan} or \code{fread}}
}
\value{
If df == TRUE, a \code{data.frame}, else a \code{list}
}
\description{
Convenience wrapper to import typical output from CWB tools.
The expected file format is tab delimited, only containing data, no quotes,
missing values, comments or header.
}
\details{
This is just a wrapper around \code{scan} that behaves similarly to
\code{read.table}.
The main difference is that \code{wc -l} is run on the file before
in order to get the number of lines to pass to \code{scan}.
This reduces memory overhead substantially and can also be a bit faster.
}
\examples{
\dontrun{
path <- "brown_word_per_id.txt"
out <- read_freqs(path, list(f = 0L, type = "", text_id = ""))
}
}
\seealso{
\code{\link[=scan]{scan()}}
}
